requirements.txt
streamlit
chromadb
python-dotenv
langchain
langchain-community
langchain-huggingface
langchain-groq
sentence-transformers